+++ packages/dart/CHANGELOG.md
@@ -5,6 +5,7 @@
 - **BREAKING CHANGES**: `strToNumberHash` has renamed to `numberHash`
 - `md5Hash`, `sha1Hash`, `sha256Hash`: Add an encoding option for hash functions
 - Add `sortNumeric` method
+- Add `sha512Hash` method
 
 ## 1.1.12 (2026-03-31)